Move db to db module
