+++ app/src/main/java/io/krugosvet/dailydish/android/architecture/injection/Koin.kt
@@ -2,12 +2,8 @@
 
 package io.krugosvet.dailydish.android.architecture.injection
 
-import android.content.Context
 import androidx.fragment.app.Fragment
-import androidx.room.Room
 import io.krugosvet.dailydish.android.architecture.view.GenericBaseActivity
-import io.krugosvet.dailydish.android.db.DatabaseService
-import io.krugosvet.dailydish.android.db.meal.MealDao
 import io.krugosvet.dailydish.android.repository.Meal
 import io.krugosvet.dailydish.android.repository.MealRepository
 import io.krugosvet.dailydish.android.screen.addMeal.viewmodel.AddMealViewModel
@@ -42,20 +38,6 @@ val module = module {
 
   viewModel<AddMealViewModel>()
   viewModel<MealListViewModel>()
-}
-
-val dbModule = module {
-
-  single<DatabaseService> {
-    Room
-      .databaseBuilder(get<Context>(), DatabaseService::class.java, "database")
-      .build()
-  }
-
-  single<MealDao> {
-    get<DatabaseService>().mealDao
-  }
-
 }
 
 inline fun <reified T : Any> GenericBaseActivity.activityInject() =
