+++ db/build.gradle
@@ -35,4 +35,8 @@ dependencies {
     implementation "androidx.room:room-runtime:$room_version"
     implementation "androidx.room:room-ktx:$room_version"
     kapt "androidx.room:room-compiler:$room_version"
+
+    // Koin
+
+    implementation "org.koin:koin-core:$koin_version"
 }
