+++ db/src/main/java/io/krugosvet/dailydish/DatabaseService.kt
@@ -1,9 +1,9 @@
-package io.krugosvet.dailydish.android.db
+package io.krugosvet.dailydish
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import io.krugosvet.dailydish.android.db.meal.MealDao
-import io.krugosvet.dailydish.android.db.meal.MealEntity
+import io.krugosvet.dailydish.meal.MealDao
+import io.krugosvet.dailydish.meal.MealEntity
 
 @Database(entities = [MealEntity::class], version = 1, exportSchema = false)
 abstract class DatabaseService: RoomDatabase() {
