+++ db/src/main/java/io/krugosvet/dailydish/meal/MealDao.kt
@@ -1,4 +1,4 @@
-package io.krugosvet.dailydish.android.db.meal
+package io.krugosvet.dailydish.meal
 
 import androidx.room.*
 import kotlinx.coroutines.flow.Flow
